Generate and test PIE packages (#31)
* Support PIE package distribution

Support distributing the extension as a PIE package
on Packagist.

Add a CD pipeline for validating that the package can
be put on a local Composer repo, then built and installed
from the repo using the pie command. The installed
package is then run against the same integration tests
as the dev pipeline.

Rework Makefile.frag so that the extension can
be built with the standard set of commands that
the pie command runs:
- phpize
- ./configure
- make
- make install

This also requires loading the valkey-glide
sub-module, generating protobufs, running the
FFI build internally within the Makefile.

Create shared setup action for running PHP tests.

Signed-off-by: James Duong <duong.james@gmail.com>